refactor: extract labels/summary to pr-triage-labels.mjs, fix rate limit handling
- Extract label application and summary output to pr-triage-labels.mjs (114 lines)
- Main orchestrator now 268 lines (was 497, under 400-line limit)
- Handle GitHub 403 rate limits in retry logic (x-ratelimit-remaining header)
- Reduce file-fetch concurrency from 10 to 5 to conserve rate limit budget
- Update workflow sparse-checkout to include new module
